ilasm: Somewhat properly support !! and ! in type signatures.
This remains a giant hack. Ideally, Cecil should allow use to create
some sort of GenericParameterReference with no owner, but this is not
possible at the moment, making it virtually impossible to be compatible
with Microsoft's behavior here.

A side-effect is that our implementation is 'safer' in the sense that
generic parameters/arguments can only be referred to if they actually
exist.