Support VFP v3 variants of ARM_FLDMD and ARM_FSTMD.
Previously, we set the D bit of these instructions to Fd & 1. This
was incorrect. The D bit should only be set if we're referring to
the extra double-precision registers in VFP v3.

Take for example the instruction:

    vstmia r0, {d18-d20}

This is encoded by the GNU assembler to:

    1110 110 0 1 1 0 0 0000 0010 1011 00000110
                 |          |
                 D          Vd

The Vd is set to 2, i.e. register d2. However, we want to refer to
the d18 register, which is why the D bit is set. The CPU concats
the D bit with the four Vd bits to get 0b10010, which is 18.

However, if we do:

    vstmia r0, {d2-d4}

We get:

    1110 110 0 1 0 0 0 0000 0010 1011 00000110
                 |          |
                 D          Vd

The only difference is that the D bit is not set. This is because
the four Vd bits are enough to refer to the 16 base double-precision
registers in VFP.