Move the unaligned 64-bit CAS hack to atomic.c.
Since we now rely on InterlockedCompareExchange64 () in quite a few
places that involve managed references, it's better to just put this
in one place.

This moves the assembly for ARM into a separate function since a
naked function can only (safely) contain asm statements.

Also added a HACK note so we don't forget about it.